chore: refactor storage check to state
