+++ src/agent.py
@@ -8,10 +8,7 @@ import typing
 import ops
 
 import jenkins
-from state import AGENT_RELATION, DEPRECATED_AGENT_RELATION, AgentMeta, State
-
-if typing.TYPE_CHECKING:
-    from charm import JenkinsK8sOperatorCharm  # pragma: no cover
+from state import AGENT_RELATION, DEPRECATED_AGENT_RELATION, JENKINS_SERVICE_NAME, AgentMeta, State
 
 logger = logging.getLogger(__name__)
 
@@ -31,7 +28,7 @@ class AgentRelationData(typing.TypedDict):
 class Observer(ops.Object):
     """The Jenkins agent relation observer."""
 
-    def __init__(self, charm: "JenkinsK8sOperatorCharm", state: State):
+    def __init__(self, charm: ops.CharmBase, state: State):
         """Initialize the observer and register event handlers.
 
         Args:
@@ -63,8 +60,8 @@ class Observer(ops.Object):
         Args:
             event: The event fired from an agent joining the relationship.
         """
-        container = self.charm.unit.get_container(self.state.jenkins_service_name)
-        if not container.can_connect() or not self.charm.is_storage_ready:
+        container = self.charm.unit.get_container(JENKINS_SERVICE_NAME)
+        if not container.can_connect() or not self.state.is_storage_ready:
             return
         # The relation is joined, it cannot be None, hence the type casting.
         deprecated_agent_relation_meta = typing.cast(
@@ -102,8 +99,8 @@ class Observer(ops.Object):
         Args:
             event: The event fired from an agent joining the relationship.
         """
-        container = self.charm.unit.get_container(self.state.jenkins_service_name)
-        if not container.can_connect() or not self.charm.is_storage_ready:
+        container = self.charm.unit.get_container(JENKINS_SERVICE_NAME)
+        if not container.can_connect() or not self.state.is_storage_ready:
             return
         # The relation is joined, it cannot be None, hence the type casting.
         agent_relation_meta = typing.cast(
@@ -145,8 +142,8 @@ class Observer(ops.Object):
             event: The event fired when a unit in deprecated agent relation is departed.
         """
         # the event unit cannot be None.
-        container = self.charm.unit.get_container(self.state.jenkins_service_name)
-        if not container.can_connect() or not self.charm.is_storage_ready:
+        container = self.charm.unit.get_container(JENKINS_SERVICE_NAME)
+        if not container.can_connect() or not self.state.is_storage_ready:
             return
 
         # The relation data is removed before this particular hook runs, making the name set by the
@@ -172,8 +169,8 @@ class Observer(ops.Object):
             event: The event fired when a unit in agent relation is departed.
         """
         # the event unit cannot be None.
-        container = self.charm.unit.get_container(self.state.jenkins_service_name)
-        if not container.can_connect() or not self.charm.is_storage_ready:
+        container = self.charm.unit.get_container(JENKINS_SERVICE_NAME)
+        if not container.can_connect() or not self.state.is_storage_ready:
             return
 
         # The relation data is removed before this particular hook runs, making the name set by the
